fix: prevents panic when payload includes null values
When a pipeline doesn not specify a source configuration for a resource,
or params/get_params on get or put steps, the corresponding payload
parameter will be set to null, not ommitted from the payload. This
change fixes the json processing logic to prevent panicing when null
values are encountered. It also modifies to the existing to behavior to
execute any user defined Validate methods regardless of whether a value
is defined, omitted, or set to null. This gives the user a chance to
populate defaults or define correct failure behavior in the event that
a value is passed as null.